chore: Add ExampleGenerator interface
